+++ packages/rspack/README.md
@@ -1,0 +1,3 @@
+# @workshop/rspack
+
+Common Rspack config to be reused in all frontends and workshop exercises
