+++ packages/rspack/src/index.js
@@ -1,0 +1,5 @@
+const rspackDevConfig = require('./rspack.dev.config');
+const rspackProdConfig = require('./rspack.prod.config');
+
+module.exports.rspackDevConfig = rspackDevConfig;
+module.exports.rspackProdConfig = rspackProdConfig;
