+++ packages/rspack/src/rspack.prod.config.js
@@ -1,12 +1,9 @@
-const path = require("path");
-
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const dotEnv = require("dotenv-webpack");
-const webpackCommonConfig = require("./webpack.config");
 const { merge } = require('webpack-merge');
 const TerserPlugin = require('terser-webpack-plugin');
 
-module.exports = merge(webpackCommonConfig, {
+const rspackCommonConfig = require('./rspack.config');
+
+module.exports = merge(rspackCommonConfig, {
   mode: 'production',
   devtool: 'source-map',
   optimization: {
