--- packages/webpack/README.md
@@ -1,3 +1,0 @@
-# @workshop/webpack
-
-Common Webpack config to be reused in all frontends and workshop exercises
