--- packages/webpack/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@workshop/webpack",
-  "version": "1.0.0",
-  "main": "src/index.js",
-  "devDependencies": {
-    "@pmmmwh/react-refresh-webpack-plugin": "0.5.15",
-    "css-loader": "7.1.2",
-    "dotenv-webpack": "^8.1.0",
-    "html-webpack-plugin": "5.6.0",
-    "webpack-cli": "^5.0.0",
-    "webpack-dev-server": "^5.0.4",
-    "webpack-merge": "6.0.1",
-    "webpack-plugin-serve": "^1.2.0"
-  }
-}
