--- packages/webpack/src/index.js
@@ -1,5 +1,0 @@
-const webpackDevConfig = require('./webpack.dev.config');
-const webpackProdConfig = require('./webpack.prod.config');
-
-module.exports.webpackDevConfig = webpackDevConfig;
-module.exports.webpackProdConfig = webpackProdConfig;
